Fix building libbinder against musl
musl libc uses the posix definition of msg_iovlen as int.  Cast the
input size_t to int to avoid a tidy failure.

Bug: 190084016
Test: m USE_HOST_MUSL=true host-native

+++ libs/binder/RpcTransportRaw.cpp
@@ -74,7 +74,7 @@ public:
         while (true) {
             msghdr msg{
                     .msg_iov = iovs,
-                    .msg_iovlen = niovs,
+                    .msg_iovlen = static_cast<int>(niovs), // posix uses int, glibc uses size_t
             };
             ssize_t processSize =
                     TEMP_FAILURE_RETRY(sendOrReceiveFun(mSocket.get(), &msg, MSG_NOSIGNAL));
